cloud_functions/load-v2-tournament-data-2/requirements.txt
numerapi==2.8.1
pandas==1.3.2
google-cloud-storage==1.18.1
google-cloud-bigquery==1.21.0
google-cloud-pubsub==2.8.0
